+++ include/pan_tilt/pan_tilt_node.h
@@ -2,10 +2,10 @@
 #define PAN_TILT_NODE_H_
 /* Pan/Tilt Processing Node. Takes demand for Pan/Tilt position and request the servo positions.
  * It can handle a number of pan tilt device. For example index 0 may be for a head/camera and
- * index 1 could be for a LIDAR device.
+ * index 1 could be for a second device.
  */
 #include <ros/ros.h>
-#include <servo_msgs/pan_tilt.h>
+#include <sensor_msgs/JointState.h>
 #include <dynamic_reconfigure/server.h>
 #include <pan_tilt/PanTiltConfig.h>
 
@@ -20,13 +20,13 @@ public:
     void reconfCallback(pan_tilt::PanTiltConfig &config, uint32_t level);
 
 private:
-    void panTilt0CB(const servo_msgs::pan_tilt& pan_tilt);    
-    void panTilt1CB(const servo_msgs::pan_tilt& pan_tilt);    
-    void movePanTilt(const servo_msgs::pan_tilt& pan_tilt, int pan_trim, int tilt_trim, int index);
+    void panTiltCB(const sensor_msgs::JointState& joint);
+    void movePanTilt(int pan_value, int tilt_value, int pan_trim, int tilt_trim, int index);
 	int checkMaxMin(int current_value, int max, int min);
+    double signedRadianToServoDegrees(double rad, bool flip_rotation);
 
     ros::NodeHandle n_;
-    ros::Subscriber pan_tilt_sub_[NUMBER_OF_PAN_TILT_DEVICES];
+    ros::Subscriber joint_state_sub_;
     ros::Publisher servo_array_pub_;
     
     // Used for dynamic reconfiguration
@@ -38,8 +38,10 @@ private:
     bool first_index1_msg_received_;
     
     // Used to store last position (not including trim value)
-    servo_msgs::pan_tilt index0_pan_tilt_;
-    servo_msgs::pan_tilt index1_pan_tilt_;
+    int index0_pan_;
+    int index0_tilt_;
+    int index1_pan_;
+    int index1_tilt_;
 
     // Configuration parameters
     int pan_servo_[NUMBER_OF_PAN_TILT_DEVICES];     // Maps a servo to a pan device
@@ -48,6 +50,10 @@ private:
     int pan_min_[NUMBER_OF_PAN_TILT_DEVICES];       // Minimum range for the pan servo
     int tilt_max_[NUMBER_OF_PAN_TILT_DEVICES];      // Maximum range for the tilt servo 
     int tilt_min_[NUMBER_OF_PAN_TILT_DEVICES];      // Minimum range for the tilt servo
+    bool pan_flip_rotation_[NUMBER_OF_PAN_TILT_DEVICES]; // true if the pan servo rotation should be flipped
+    bool tilt_flip_rotation_[NUMBER_OF_PAN_TILT_DEVICES]; // true if the tilt servo rotation should be flipped
+    std::string pan_joint_names_[NUMBER_OF_PAN_TILT_DEVICES];   // Names of the pan joints
+    std::string tilt_joint_names_[NUMBER_OF_PAN_TILT_DEVICES];  // Names of the tilt joinst
 };
 
 #endif // PAN_TILT_NODE_H_